requirements.txt
pandas
numpy
scikit-learn
streamlit
requests
Pillow
pyngrok
